% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{anonymous_login}
\alias{anonymous_login}
\title{Anonymous Login (Legacy)}
\usage{
anonymous_login(project_api)
}
\arguments{
\item{project_api}{The Firebase Project API key}
}
\value{
Authentication response
}
\description{
Legacy anonymous login. Use auth_anonymous() instead.
}
