% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_get_user}
\alias{auth_get_user}
\title{Get Current User Data}
\usage{
auth_get_user(conn = NULL, id_token = NULL, api_key = NULL)
}
\arguments{
\item{conn}{Firebase connection object with valid token}

\item{id_token}{ID token (used if conn doesn't have token)}

\item{api_key}{Firebase API key}
}
\value{
User account information
}
\description{
Retrieves account info for the currently authenticated user.
}
\examples{
\dontrun{
user_info <- auth_get_user(conn)
}
}
