% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{base64url_encode}
\alias{base64url_encode}
\title{Base64URL Encode}
\usage{
base64url_encode(data)
}
\arguments{
\item{data}{Raw bytes or string to encode}
}
\value{
Base64URL encoded string
}
\description{
Base64URL Encode
}
\keyword{internal}
