% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{batchGetDocuments}
\alias{batchGetDocuments}
\title{The firestore batch get function}
\usage{
batchGetDocuments(projectID, documents, databaseID = "(default)",
  mask = NULL, transaction = NULL, newTransaction = NULL,
  readTime = NULL, token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{documents}{A vector consists of documents names in the format projects/(projectId)/databases/(databaseId)/documents/(document_path) {string}}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{mask}{The fields to return. If not set, returns all fields. {string}}

\item{transaction}{Reads documents in a transaction {string}}

\item{newTransaction}{Starts a new transaction and reads the documents.
Defaults to a read-only transaction. The new transaction ID will be returned as the first response in the stream.{string}}

\item{readTime}{Reads documents as they were at the given time. This may not be older than 60 seconds. {string}}

\item{token}{The user access token that can be retrieved with the auth() function.
An OAuth 2.0 access token is required for listCollectionIDs. {string}}
}
\value{
A HTTP response that contains the documents, transaction, readTime and missing documents
}
\description{
Get multiple documents by names
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
