% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{beginTransaction}
\alias{beginTransaction}
\title{The firestore beginTransaction function}
\usage{
beginTransaction(projectID, options, databaseID = "(default)",
  token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{options}{options for the transactions {string}}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{token}{The user access token that can be retrieved with the auth() function. Required when the database rules specify the need for user authentications. {string}}
}
\value{
http response with the transaction number
}
\description{
Begin a transaction
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
