% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{build_firestore_collection_url}
\alias{build_firestore_collection_url}
\title{Build Firestore Collection URL}
\usage{
build_firestore_collection_url(project_id, collection)
}
\description{
Build Firestore Collection URL
}
\keyword{internal}
