% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{build_firestore_url}
\alias{build_firestore_url}
\title{Build Firebase Firestore URL}
\usage{
build_firestore_url(project_id)
}
\arguments{
\item{project_id}{Firebase project ID}
}
\value{
The Firestore base URL
}
\description{
Build Firebase Firestore URL
}
\keyword{internal}
