% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{build_rtdb_path_url}
\alias{build_rtdb_path_url}
\title{Build Realtime Database Path URL}
\usage{
build_rtdb_path_url(base_url, path)
}
\arguments{
\item{base_url}{Database base URL}

\item{path}{Database path}
}
\value{
Complete URL with .json suffix
}
\description{
Build Realtime Database Path URL
}
\keyword{internal}
