% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{build_rtdb_url}
\alias{build_rtdb_url}
\title{Build Firebase Realtime Database URL}
\usage{
build_rtdb_url(project_id = NULL, database_url = NULL)
}
\arguments{
\item{project_id}{Firebase project ID}

\item{database_url}{Exact database URL (recommended for regional databases)}
}
\value{
The database base URL
}
\description{
Build Firebase Realtime Database URL
}
\keyword{internal}
