% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{build_storage_upload_url}
\alias{build_storage_upload_url}
\title{Build a Cloud Storage Media Upload URL}
\usage{
build_storage_upload_url(bucket, object, predefined_acl = NULL)
}
\arguments{
\item{bucket}{Storage bucket name}

\item{object}{Object path}

\item{predefined_acl}{Optional predefined object ACL. Leave NULL for buckets
that use uniform bucket-level access.}
}
\value{
Complete media upload URL
}
\description{
Build a Cloud Storage Media Upload URL
}
\keyword{internal}
