% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{build_storage_url}
\alias{build_storage_url}
\title{Build Storage API URL}
\usage{
build_storage_url(bucket, object = NULL, upload = FALSE)
}
\arguments{
\item{bucket}{Bucket name}

\item{object}{Object path (optional)}

\item{upload}{Whether this is an upload URL}
}
\value{
Complete storage URL
}
\description{
Build Storage API URL
}
\keyword{internal}
