% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{build_token_url}
\alias{build_token_url}
\title{Build Token Refresh URL}
\usage{
build_token_url(api_key)
}
\arguments{
\item{api_key}{Firebase API key}
}
\value{
Complete URL for token refresh
}
\description{
Build Token Refresh URL
}
\keyword{internal}
