% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_conversion}
\alias{class_conversion}
\alias{classConversion}
\title{Convert S4 Class to Base64}
\usage{
class_conversion(x)

classConversion(x)
}
\arguments{
\item{x}{The S4 class object}
}
\value{
A list containing the base64-encoded binary data
}
\description{
Internal function that serializes S4 class objects to base64-encoded
binary format for storage in Firebase.
}
\keyword{internal}
