% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{class_restore}
\alias{class_restore}
\title{Restore S4 Class from Base64}
\usage{
class_restore(data)
}
\arguments{
\item{data}{List containing base64Set field}
}
\value{
The restored S4 object
}
\description{
Restores an S4 class object from base64-encoded binary data.
}
\keyword{internal}
