% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_path}
\alias{clean_path}
\title{Clean Firebase Path}
\usage{
clean_path(path)
}
\arguments{
\item{path}{Path to clean}
}
\value{
Cleaned path
}
\description{
Removes leading/trailing slashes and normalizes path.
}
\keyword{internal}
