% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_paths}
\alias{combine_paths}
\title{Combine Paths}
\usage{
combine_paths(...)
}
\arguments{
\item{...}{Path segments to combine}
}
\value{
Combined path
}
\description{
Combines path segments, handling slashes properly.
}
\keyword{internal}
