% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{conditionMessage.firebase_error}
\alias{conditionMessage.firebase_error}
\title{Condition message for firebase_error}
\usage{
\method{conditionMessage}{firebase_error}(c)
}
\arguments{
\item{c}{A firebase_error condition}
}
\value{
The error message as a character string
}
\description{
Condition message for firebase_error
}
\keyword{internal}
