% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{createIndex}
\alias{createIndex}
\title{The firestore create index function}
\usage{
createIndex(projectID, index, databaseID = "(default)", token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{index}{representation of an index {string}}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{token}{The user access token that can be retrieved with the auth() function.
An OAuth 2.0 access token is required for listCollectionIDs. {string}}
}
\value{
A response that contains the status of the operation
}
\description{
Create an index on firestore
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
