% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{dataBackup}
\alias{dataBackup}
\title{The firebase database backup function:}
\usage{
dataBackup(projectURL, secretKey = "prompt", fileName)
}
\arguments{
\item{secretKey}{The firebase secret key, which can be found in the Config/ Service Accounts/ Database secrets firebase page. {string}}

\item{fileName}{The output file name. Can be any string with .json format {string}}

\item{projectUrl}{The Firebase Project Url {string}}
}
\value{
Returns either a warning or the backup file name.
}
\description{
The backup functionality allows to download the whole database into a .json file (which can later be uploaded in the firebase console to do a restore of the DB). Generally this function may allow to save costs by not relying on the Firebase automatic backup function that is only available with the Firebase Blaze premium payment contract.
}
