% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{decode}
\alias{decode}
\title{Decode a Firestore document to R variable}
\usage{
decode(response, parse = TRUE)
}
\arguments{
\item{response}{HTTP response including a Firestore document}

\item{parse}{whether to parse the response first, default is true}
}
\value{
R variable
}
\description{
Convert string which conforms to Firestore document format to R variable
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
