% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{delete}
\alias{delete}
\title{Delete Data (Legacy)}
\usage{
delete(x, projectURL, directory = "main", token = "none")
}
\arguments{
\item{x}{Data (unused, for signature compatibility)}

\item{projectURL}{Firebase database URL}

\item{directory}{Database path}

\item{token}{Authentication token}
}
\value{
Path
}
\description{
Legacy delete function. Use rtdb_delete() instead.
}
