% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{delete_storage}
\alias{delete_storage}
\title{The list function for firebase storage:}
\usage{
delete_storage(bucket_name, object_name, web_client_id, web_client_secret)
}
\arguments{
\item{bucket_name}{The name of your storage bucket. {string}}

\item{object_name}{The name of the file that you want to delete from the bucket. {string}}

\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}
}
\value{
Returns NULL when successful.
}
\description{
fireData::list_storage lists all files in the firebase storage bucket.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
