% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{deleteDocument}
\alias{deleteDocument}
\title{The firestore deleteDocument function}
\usage{
deleteDocument(projectID, documentPath, databaseID = "(default)",
  token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{documentPath}{path for the document to be deleted {string}}

\item{databaseID}{The database under which document will be deleted {string}}

\item{token}{The user access token that can be retrieved with the auth() function. Required when the database rules specify the need for user authentications. {string}}
}
\value{
returns empty when the operation is successful, otherwise a http response with error
}
\description{
The function allows to delete a document on firestore databases
}
\examples{
\dontrun{
response <- deleteDocument("gsoc2018-d05d8", documentPath = "this/trythis")
}
}
\author{
Jiasheng Zhu
}
