% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{deleteIndex}
\alias{deleteIndex}
\title{The firestore delete index function}
\usage{
deleteIndex(projectID, indexID, databaseID = "(default)", token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{indexID}{ID of the index to delete {string}}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{token}{The user access token {string}}
}
\value{
If successful, the response body will be empty
}
\description{
Delete an index on firestore
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
