% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{download}
\alias{download}
\title{Download Data (Legacy)}
\usage{
download(
  projectURL,
  fileName,
  secretKey = "none",
  token = "none",
  isClass = FALSE
)
}
\arguments{
\item{projectURL}{Firebase database URL}

\item{fileName}{Database path}

\item{secretKey}{Admin secret key}

\item{token}{Authentication token}

\item{isClass}{Whether data is S4 class}
}
\value{
Downloaded data
}
\description{
Legacy download function. Use rtdb_get() instead.
}
