% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{encode}
\alias{encode}
\title{Encode a R variable to Firestore document}
\usage{
encode(value)
}
\arguments{
\item{value}{The data frame to be converted}
}
\value{
string representing Firestore document
}
\description{
Convert any R variable to string which conforms to Firestore document format
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
