% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{ENV_VARS}
\alias{ENV_VARS}
\title{Environment Variable Names}
\format{
An object of class \code{list} of length 8.
}
\usage{
ENV_VARS
}
\description{
Environment Variable Names
}
\keyword{internal}
