% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_conversion}
\alias{file_conversion}
\alias{fileConversion}
\title{Convert Data for Firebase Storage}
\usage{
file_conversion(x)

fileConversion(x)
}
\arguments{
\item{x}{The input data (data.frame, list, or S4 object)}
}
\value{
Converted data suitable for JSON serialization
}
\description{
Internal function that converts R objects to Firebase-compatible format.
S4 class objects are converted to base64-encoded binary.
}
\keyword{internal}
