% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{fileConversion}
\alias{fileConversion}
\title{Data conversion function}
\usage{
fileConversion(x)
}
\arguments{
\item{x}{the input file.}
}
\value{
returns optionally reformatted data.
}
\description{
The internal data conversion function to bring data in the right json format. In case the uploaded file is a s4 class object, the object is converted to a binary s4 object.
}
