% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{find_config_file}
\alias{find_config_file}
\title{Find Configuration File}
\usage{
find_config_file()
}
\value{
Path to config file or NULL if not found
}
\description{
Searches for configuration file in standard locations:
\enumerate{
\item .firedata.yml in current directory
\item .firedata.yml in home directory
\item ~/.firedata/config.yml
}
}
\keyword{internal}
