% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{firebase_auth_error}
\alias{firebase_auth_error}
\title{Firebase Authentication Error}
\usage{
firebase_auth_error(message, code = NULL, ...)
}
\arguments{
\item{message}{Error message}

\item{code}{Firebase auth error code}

\item{...}{Additional arguments passed to firebase_error}
}
\value{
A condition object of class firebase_auth_error
}
\description{
Firebase Authentication Error
}
\keyword{internal}
