% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{firebase_delete}
\alias{firebase_delete}
\title{Firebase DELETE Request}
\usage{
firebase_delete(url, ...)
}
\arguments{
\item{url}{Request URL}

\item{...}{Additional arguments passed to firebase_request}
}
\value{
Parsed response
}
\description{
Firebase DELETE Request
}
\keyword{internal}
