% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{firebase_get}
\alias{firebase_get}
\title{Firebase GET Request}
\usage{
firebase_get(url, ...)
}
\arguments{
\item{url}{Request URL}

\item{...}{Additional arguments passed to firebase_request}
}
\value{
Parsed response
}
\description{
Firebase GET Request
}
\keyword{internal}
