% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{firebase_network_error}
\alias{firebase_network_error}
\title{Firebase Network Error}
\usage{
firebase_network_error(message, ...)
}
\arguments{
\item{message}{Error message}

\item{...}{Additional arguments passed to firebase_error}
}
\value{
A condition object of class firebase_network_error
}
\description{
Firebase Network Error
}
\keyword{internal}
