% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{firebase_not_found_error}
\alias{firebase_not_found_error}
\title{Firebase Not Found Error}
\usage{
firebase_not_found_error(message, path = NULL, ...)
}
\arguments{
\item{message}{Error message}

\item{path}{The path that was not found}

\item{...}{Additional arguments passed to firebase_error}
}
\value{
A condition object of class firebase_not_found_error
}
\description{
Firebase Not Found Error
}
\keyword{internal}
