% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{firebase_patch}
\alias{firebase_patch}
\title{Firebase PATCH Request}
\usage{
firebase_patch(url, body = NULL, ...)
}
\arguments{
\item{url}{Request URL}

\item{body}{Request body}

\item{...}{Additional arguments passed to firebase_request}
}
\value{
Parsed response
}
\description{
Firebase PATCH Request
}
\keyword{internal}
