% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{firebase_permission_error}
\alias{firebase_permission_error}
\title{Firebase Permission Denied Error}
\usage{
firebase_permission_error(message, path = NULL, ...)
}
\arguments{
\item{message}{Error message}

\item{path}{The path that was denied}

\item{...}{Additional arguments passed to firebase_error}
}
\value{
A condition object of class firebase_permission_error
}
\description{
Firebase Permission Denied Error
}
\keyword{internal}
