% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{firebase_set_token}
\alias{firebase_set_token}
\title{Update Connection Token}
\usage{
firebase_set_token(conn, token)
}
\arguments{
\item{conn}{Firebase connection object}

\item{token}{New token (FirebaseToken object or string)}
}
\value{
Updated connection object
}
\description{
Updates the authentication token on a connection.
}
\examples{
\dontrun{
conn <- firebase_connect(project_id = "my-project", api_key = "...")
auth_result <- auth_sign_in(conn, email, password)
conn <- firebase_set_token(conn, auth_result)
}
}
