% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{firebase_validation_error}
\alias{firebase_validation_error}
\title{Firebase Validation Error}
\usage{
firebase_validation_error(message, field = NULL, ...)
}
\arguments{
\item{message}{Error message}

\item{field}{The field that failed validation}

\item{...}{Additional arguments passed to firebase_error}
}
\value{
A condition object of class firebase_validation_error
}
\description{
Firebase Validation Error
}
\keyword{internal}
