% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{FireData_auth}
\alias{FireData_auth}
\title{FireData_auth()}
\usage{
FireData_auth(projectAPI, email, password)
}
\arguments{
\item{projectAPI}{The Firebase Project API {string}}

\item{email}{The user email {string}}

\item{password}{The user password {string}}
}
\value{
Returns the content of the firebase API request, such as the state of registration, idToken, and validity of the user password.
}
