% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{FireData_createUser}
\alias{FireData_createUser}
\title{FireData_createUser()}
\usage{
FireData_createUser(projectAPI, email, password)
}
\arguments{
\item{projectAPI}{The Firebase Project API {string}}

\item{email}{The user email {string}}

\item{password}{The user password {string}}
}
\value{
Registers a new user and returns the status.
}
