% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{fireData_download}
\alias{fireData_download}
\title{FireData_download()}
\usage{
fireData_download(x, projectURL, directory)
}
\arguments{
\item{x}{A data.frame or data.table}
}
\value{
showing shapiro.test output of the data.frame
}
