% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{firestore_decode_fields}
\alias{firestore_decode_fields}
\title{Decode Firestore Fields to R List}
\usage{
firestore_decode_fields(fields)
}
\description{
Decode Firestore Fields to R List
}
\keyword{internal}
