% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{firestore_decode_value}
\alias{firestore_decode_value}
\title{Decode Firestore Value to R Value}
\usage{
firestore_decode_value(value)
}
\description{
Decode Firestore Value to R Value
}
\keyword{internal}
