% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{firestore_encode_fields}
\alias{firestore_encode_fields}
\title{Encode R List to Firestore Fields}
\usage{
firestore_encode_fields(data)
}
\description{
Encode R List to Firestore Fields
}
\keyword{internal}
