% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{firestore_encode_value}
\alias{firestore_encode_value}
\title{Encode R Value to Firestore Value}
\usage{
firestore_encode_value(value)
}
\description{
Encode R Value to Firestore Value
}
\keyword{internal}
