% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{firestore_parse_document}
\alias{firestore_parse_document}
\title{Parse Firestore Document Response}
\usage{
firestore_parse_document(doc)
}
\description{
Parse Firestore Document Response
}
\keyword{internal}
