% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_bytes}
\alias{format_bytes}
\title{Format Bytes for Display}
\usage{
format_bytes(bytes)
}
\arguments{
\item{bytes}{Number of bytes}
}
\value{
Human-readable string
}
\description{
Format Bytes for Display
}
\keyword{internal}
