% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{fs_limit}
\alias{fs_limit}
\title{Limit Firestore Query Results}
\usage{
fs_limit(query, n)
}
\arguments{
\item{query}{A firestore_query object}

\item{n}{Maximum number of documents to return}
}
\value{
Updated firestore_query object
}
\description{
Limits the number of documents returned by a query.
}
