% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{fs_offset}
\alias{fs_offset}
\title{Offset Firestore Query Results}
\usage{
fs_offset(query, n)
}
\arguments{
\item{query}{A firestore_query object}

\item{n}{Number of documents to skip}
}
\value{
Updated firestore_query object
}
\description{
Skips a number of documents before returning results.
}
