% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{fs_select}
\alias{fs_select}
\title{Select Fields in Firestore Query}
\usage{
fs_select(query, ...)
}
\arguments{
\item{query}{A firestore_query object}

\item{...}{Field names to select}
}
\value{
Updated firestore_query object
}
\description{
Specifies which fields to return in query results.
}
