% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_id}
\alias{generate_id}
\title{Generate Random ID}
\usage{
generate_id(length = 20)
}
\arguments{
\item{length}{Length of ID (default: 20)}
}
\value{
Random ID string
}
\description{
Generates a random ID similar to Firebase push IDs.
}
\keyword{internal}
