% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{generateDocument}
\alias{generateDocument}
\title{Generate Firestore document}
\usage{
generateDocument(data, name = "none")
}
\arguments{
\item{data}{The data frame to be converted {data.frame}}

\item{name}{an optional name field can be added to the document}

\item{max_digits}{The maximum digits to keep in the convertion, default is NA, (all)}

\item{auto_unbox}{automatically \code{\link{unbox}} all atomic vectors of length 1. It is usually safer to avoid this and instead use the \code{\link{unbox}} function to unbox individual elements.
An exception is that objects of class \code{AsIs} (i.e. wrapped in \code{I()}) are not automatically unboxed. This is a way to mark single values as length-1 arrays.}
}
\value{
converted json string
}
\description{
Convert data frame to a Firestore document
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
