% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_file_value}
\alias{get_config_file_value}
\title{Get Value from Config File}
\usage{
get_config_file_value(key, profile = "default")
}
\arguments{
\item{key}{Configuration key}

\item{profile}{Configuration profile}
}
\value{
Value or NULL
}
\description{
Get Value from Config File
}
\keyword{internal}
