% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_connection_token}
\alias{get_connection_token}
\title{Get Token from Connection}
\usage{
get_connection_token(conn, require_auth = FALSE)
}
\arguments{
\item{conn}{Firebase connection object}

\item{require_auth}{Whether to error if no token available}
}
\value{
Token string or NULL
}
\description{
Extracts or resolves an authentication token from the connection.
}
\keyword{internal}
