% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-links.R
\name{get_dynamic_link}
\alias{get_dynamic_link}
\title{Create Dynamic Link (Legacy)}
\usage{
get_dynamic_link(
  project_api,
  domain,
  link,
  short = TRUE,
  social_title = "",
  social_description = "",
  social_image_link = ""
)
}
\arguments{
\item{project_api}{Firebase API key}

\item{domain}{Dynamic Links domain}

\item{link}{URL to shorten}

\item{short}{Whether to create short link}

\item{social_title}{Social preview title}

\item{social_description}{Social preview description}

\item{social_image_link}{Social preview image}
}
\value{
This function always raises an error because the upstream service has
been shut down.
}
\description{
Legacy function. Use dynlink_create() instead.
}
