% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{get_oauth_token}
\alias{get_oauth_token}
\title{Get OAuth2 Token for Google APIs}
\usage{
get_oauth_token(client_id, client_secret, scopes, cache = FALSE)
}
\arguments{
\item{client_id}{OAuth2 client ID}

\item{client_secret}{OAuth2 client secret}

\item{scopes}{OAuth2 scopes to request}

\item{cache}{Whether to cache the token}
}
\value{
An httr OAuth2 token object
}
\description{
Creates an OAuth2 token for accessing Google APIs.
}
\keyword{internal}
