% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{get_storage}
\alias{get_storage}
\title{Get Storage Metadata (Legacy)}
\usage{
get_storage(
  bucket_name,
  object_name,
  web_client_id = "prompt",
  web_client_secret = "prompt"
)
}
\arguments{
\item{bucket_name}{Storage bucket name}

\item{object_name}{Object name}

\item{web_client_id}{OAuth client ID}

\item{web_client_secret}{OAuth client secret}
}
\value{
Object metadata
}
\description{
Legacy function. Use storage_get_metadata() instead.
}
