% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_user_token}
\alias{get_user_token}
\title{Get User Token}
\usage{
get_user_token(user)
}
\arguments{
\item{user}{The reactiveValues object used with shiny_auth_server}
}
\value{
The ID token string or NULL if not authenticated
}
\description{
Gets the Firebase token from authenticated user in Shiny context.
}
