% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{getDocument}
\alias{getDocument}
\title{The firestore get function}
\usage{
getDocument(projectID, documentPath, databaseID = "(default)",
  token = "none", decode = TRUE)
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{documentPath}{path for the document to be deleted {string}}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{token}{The user access token that can be retrieved with the auth() function. Required when the database rules specify the need for user authentications. {string}}

\item{decode}{Whether to decode the result to R variblaes}
}
\value{
returns the data frame if successful, else return the full http response
}
\description{
Get single document
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
