% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{getIndex}
\alias{getIndex}
\title{The firestore get index function}
\usage{
getIndex(projectID, indexID, databaseID = "(default)", token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{indexID}{ID of the index to delete {string}}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{token}{The user access token {string}}
}
\value{
If successful, the response body will contain an instance of the index
}
\description{
Get an index on firestore
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
