% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{google_devstorage_read_only}
\alias{google_devstorage_read_only}
\title{Get Storage Token with Read-Only Access (Legacy)}
\usage{
google_devstorage_read_only(
  web_client_id = "prompt",
  web_client_secret = "prompt"
)
}
\arguments{
\item{web_client_id}{OAuth client ID}

\item{web_client_secret}{OAuth client secret}
}
\value{
OAuth token
}
\description{
Legacy function for OAuth token.
}
