% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{google_devstorage_read_write}
\alias{google_devstorage_read_write}
\title{The OAuth function to get read/write access to the storage:}
\usage{
google_devstorage_read_write(web_client_id = "prompt",
  web_client_secret = "prompt")
}
\arguments{
\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}
}
\value{
Returns the token data.
}
\description{
fireData::google_devstorage_read_write retrieves a token with read/write access to the storage.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
