% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{google_firestore}
\alias{google_firestore}
\title{Get Firestore Token (Legacy)}
\usage{
google_firestore(
  web_client_id = "prompt",
  web_client_secret = "prompt",
  cache = FALSE
)
}
\arguments{
\item{web_client_id}{OAuth client ID}

\item{web_client_secret}{OAuth client secret}

\item{cache}{Whether to cache token}
}
\value{
OAuth token
}
\description{
Legacy function for Firestore OAuth token.
}
