% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{google_login}
\alias{google_login}
\title{The OAuth login function for google:}
\usage{
google_login(project_api, web_client_id = "prompt",
  web_client_secret = "prompt", request_uri,
  redirect_uri = httr::oauth_callback(), return_idp_credential = TRUE,
  cache = FALSE)
}
\arguments{
\item{project_api}{The Firebase Project API {string}}

\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}

\item{request_uri}{The URI to which the IDP redirects the user back. {string}}

\item{redirect_uri}{The URI to be redirected after the authentification. {string}}

\item{return_idp_credential}{Whether to force the return of the OAuth credential on the following errors: FEDERATED_USER_ID_ALREADY_LINKED and EMAIL_EXISTS. {boolean}}

\item{cache}{Cache the tokens in the .httr-oauth file or not. {boolean}}
}
\value{
Returns the content of the firebase API request, such as the idToken, the refreshToken, and the localId.
}
\description{
fireData::google_login signs in a user with OAuth credentials.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
