% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{\%+\%}
\alias{\%+\%}
\title{String concatenation operator}
\usage{
a \%+\% b
}
\arguments{
\item{a}{First string}

\item{b}{Second string}
}
\value{
Concatenated string
}
\description{
String concatenation operator
}
\keyword{internal}
