% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\title{The firestore index function}
\usage{
index(collectionID, fields)
}
\arguments{
\item{collectionID}{The collection ID to which this index applies {string}}

\item{fields}{The fields to index {string}}
}
\value{
returns a json representation of the index
}
\description{
Create an index
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
