% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{is_authenticated}
\alias{is_authenticated}
\title{Check Authentication Status}
\usage{
is_authenticated(user)
}
\arguments{
\item{user}{The reactiveValues object used with shiny_auth_server}
}
\value{
TRUE if user is logged in, FALSE otherwise
}
\description{
Helper to check if user is authenticated in Shiny context.
}
