% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_encoded_s4}
\alias{is_encoded_s4}
\title{Check if Data is Base64-Encoded S4 Object}
\usage{
is_encoded_s4(data)
}
\arguments{
\item{data}{Data to check}
}
\value{
TRUE if data appears to be a base64-encoded S4 object
}
\description{
Check if Data is Base64-Encoded S4 Object
}
\keyword{internal}
