% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{is_error}
\alias{is_error}
\title{Check if Response Indicates Error}
\usage{
is_error(response)
}
\arguments{
\item{response}{An httr response object}
}
\value{
TRUE if error, FALSE otherwise
}
\description{
Check if Response Indicates Error
}
\keyword{internal}
