% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{is_firebase_connection}
\alias{is_firebase_connection}
\title{Check if Object is a Firebase Connection}
\usage{
is_firebase_connection(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
TRUE if x is a firebase_connection
}
\description{
Check if Object is a Firebase Connection
}
