% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{is_success}
\alias{is_success}
\title{Check if Response Indicates Success}
\usage{
is_success(response)
}
\arguments{
\item{response}{An httr response object}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Check if Response Indicates Success
}
\keyword{internal}
