% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{list_storage}
\alias{list_storage}
\title{List Storage (Legacy)}
\usage{
list_storage(
  bucket_name,
  web_client_id = "prompt",
  web_client_secret = "prompt"
)
}
\arguments{
\item{bucket_name}{Storage bucket name}

\item{web_client_id}{OAuth client ID}

\item{web_client_secret}{OAuth client secret}
}
\value{
List of objects
}
\description{
Legacy function. Use storage_list() instead.
}
