% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{listCollectionIds}
\alias{listCollectionIds}
\title{The firestore list collectionID function}
\usage{
listCollectionIds(projectID, documentPath, pageSize, databaseID = "(default)",
  pageToken = "none", token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{documentPath}{path to a document {string}}

\item{pageSize}{The maximum number of collection ids to return}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{pageToken}{A page token. Used to identify a previous list collectionIDs operation {string}}

\item{token}{The user access token that can be retrieved with the auth() function.
An OAuth 2.0 access token is required for listCollectionIDs. {string}}
}
\value{
A HTTP response that contains the collectionIDs and a nextPageToken
}
\description{
Lists all the collection IDs underneath a document
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
