% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{listDocuments}
\alias{listDocuments}
\title{The firestore list function}
\usage{
listDocuments(projectID, collectionPath, pageSize, databaseID = "(default)",
  pageToken = NULL, orderBy = NULL, showMissing = FALSE, token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{collectionPath}{path to a collection {string}}

\item{pageSize}{The maximum number of documents to return}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{pageToken}{A page token. Used to identify a previous list collectionIDs operation {string}}

\item{orderBy}{The order to sort results by, for example, priority desc, name{string}}

\item{showMissing}{If the list should show missing documents.
A missing document is a document that does not exist but has sub-documents.
These documents will be returned with a key but will not have fields, Document.create_time, or Document.update_time set.
Requests with showMissing may not specify where or orderBy.}

\item{token}{The user access token that can be retrieved with the auth() function.
An OAuth 2.0 access token is required for listCollectionIDs. {string}}
}
\value{
A HTTP response that contains the documents and a nextPageToken
}
\description{
Lists all documents underneath a collection id
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
