% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{parse_firebase_error}
\alias{parse_firebase_error}
\title{Parse Firebase API Error Response}
\usage{
parse_firebase_error(response)
}
\arguments{
\item{response}{An httr response object}
}
\value{
A firebase_error subclass based on the error type
}
\description{
Parses error responses from Firebase REST APIs and returns the appropriate
error object.
}
\keyword{internal}
