% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{parse_response}
\alias{parse_response}
\title{Parse HTTP Response}
\usage{
parse_response(response)
}
\arguments{
\item{response}{An httr response object}
}
\value{
Parsed response content
}
\description{
Parse HTTP Response
}
\keyword{internal}
