% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{patch}
\alias{patch}
\title{Patch Data (Legacy)}
\usage{
patch(x, projectURL, directory = "main", token = "none")
}
\arguments{
\item{x}{Data to update}

\item{projectURL}{Firebase database URL}

\item{directory}{Database path}

\item{token}{Authentication token}
}
\value{
Path
}
\description{
Legacy patch function. Use rtdb_update() instead.
}
