% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{patchDocument}
\alias{patchDocument}
\title{The firestore patch function}
\usage{
patchDocument(projectID, documentPath, document, databaseID = "(default)",
  token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{documentPath}{path for the document to be deleted {string}}

\item{document}{The document represented in the form firestore requires}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{token}{The user access token that can be retrieved with the auth() function. Required when the database rules specify the need for user authentications. {string}}
}
\value{
returns a http response with the document if successful
}
\description{
Patch single document
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
