% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{path_check}
\alias{path_check}
\title{path_check()}
\usage{
path_check(path)
}
\arguments{
\item{x}{the path {string}}
}
\value{
the approved and cleaned path_string
}
\description{
replaces all disallowed path symbols with a "-"
}
