% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.firebase_connection}
\alias{print.firebase_connection}
\title{Print Firebase Connection}
\usage{
\method{print}{firebase_connection}(x, ...)
}
\arguments{
\item{x}{A firebase_connection object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Firebase Connection
}
