% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{print.firebase_error}
\alias{print.firebase_error}
\title{Print method for firebase_error}
\usage{
\method{print}{firebase_error}(x, ...)
}
\arguments{
\item{x}{A firebase_error object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print method for firebase_error
}
\keyword{internal}
