% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestore.R
\name{print.firestore_query}
\alias{print.firestore_query}
\title{Print Firestore Query}
\usage{
\method{print}{firestore_query}(x, ...)
}
\arguments{
\item{x}{A firestore_query object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Firestore Query
}
