% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{print.rtdb_query}
\alias{print.rtdb_query}
\title{Print Query}
\usage{
\method{print}{rtdb_query}(x, ...)
}
\arguments{
\item{x}{Query object}

\item{...}{Ignored}
}
\value{
Invisibly returns x
}
\description{
Print Query
}
