% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{put}
\alias{put}
\title{Put Data (Legacy)}
\usage{
put(x, projectURL, directory = "main", token = "none")
}
\arguments{
\item{x}{Data to write}

\item{projectURL}{Firebase database URL}

\item{directory}{Database path}

\item{token}{Authentication token}
}
\value{
Path
}
\description{
Legacy put function. Use rtdb_set() instead.
}
