% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{query_end_at}
\alias{query_end_at}
\title{End Query at Value}
\usage{
query_end_at(query, value)
}
\arguments{
\item{query}{Query builder object}

\item{value}{Ending value (inclusive)}
}
\value{
Updated query builder
}
\description{
End Query at Value
}
