% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{query_equal_to}
\alias{query_equal_to}
\title{Filter Query by Exact Value}
\usage{
query_equal_to(query, value)
}
\arguments{
\item{query}{Query builder object}

\item{value}{Value to match exactly}
}
\value{
Updated query builder
}
\description{
Filter Query by Exact Value
}
