% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{query_execute}
\alias{query_execute}
\title{Execute Query}
\usage{
query_execute(query, token = NULL)
}
\arguments{
\item{query}{Query builder object}

\item{token}{Authentication token (optional)}
}
\value{
Query results
}
\description{
Execute Query
}
