% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{query_limit_to_first}
\alias{query_limit_to_first}
\title{Limit Query to First N Results}
\usage{
query_limit_to_first(query, n)
}
\arguments{
\item{query}{Query builder object}

\item{n}{Number of results}
}
\value{
Updated query builder
}
\description{
Limit Query to First N Results
}
