% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{query_order_by}
\alias{query_order_by}
\title{Order Query by Field}
\usage{
query_order_by(query, field)
}
\arguments{
\item{query}{Query builder object}

\item{field}{Field to order by ("$key", "$value", "$priority", or child key)}
}
\value{
Updated query builder
}
\description{
Order Query by Field
}
