% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{query_start_at}
\alias{query_start_at}
\title{Start Query at Value}
\usage{
query_start_at(query, value)
}
\arguments{
\item{query}{Query builder object}

\item{value}{Starting value (inclusive)}
}
\value{
Updated query builder
}
\description{
Start Query at Value
}
