% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{recursive_decode}
\alias{recursive_decode}
\title{Decode a Firestore document to R variable recursively}
\usage{
recursive_decode(fields)
}
\arguments{
\item{fields}{a potion of a Firestore document}
}
\value{
R variable
}
\description{
Convert Firestore document to R variable recursively
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
