% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{resetPassword}
\alias{resetPassword}
\title{FireData_resetPassword()}
\usage{
resetPassword(projectAPI, email)
}
\arguments{
\item{projectAPI}{The Firebase Project API {string}}

\item{email}{The user email {string}}
}
\value{
Sends a reset email to respective user. Returns a success or warning message.
}
