% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{response_to_df}
\alias{response_to_df}
\title{Convert Response to Data Frame}
\usage{
response_to_df(data, id_column = ".id")
}
\arguments{
\item{data}{Firebase response data}

\item{id_column}{Name of column to store record IDs (default: ".id")}
}
\value{
A data frame
}
\description{
Converts Firebase response (typically a named list of records) to a data frame.
}
\keyword{internal}
