% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{rollback}
\alias{rollback}
\title{The firestore rollback function}
\usage{
rollback(projectID, transaction, databaseID = "(default)", token = "none")
}
\arguments{
\item{projectID}{The Firestore project ID {string}}

\item{transaction}{The transaction to be rolled-back {string}}

\item{databaseID}{The database under which this operation will be performed {string}}

\item{token}{The user access token that can be retrieved with the auth() function. Required when the database rules specify the need for user authentications. {string}}
}
\value{
empty http response if successful
}
\description{
Rollback a transaction
}
\examples{
\dontrun{
}
}
\author{
Jiasheng Zhu
}
