% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_delete}
\alias{storage_delete}
\title{Delete File from Storage}
\usage{
storage_delete(conn, object_name, client_id = NULL, client_secret = NULL)
}
\arguments{
\item{conn}{Firebase connection object}

\item{object_name}{Name/path of the object to delete}

\item{client_id}{OAuth client ID}

\item{client_secret}{OAuth client secret}
}
\value{
NULL on success
}
\description{
Deletes a file from Firebase Cloud Storage.
}
\examples{
\dontrun{
storage_delete(conn, "images/old-photo.jpg")
}
}
