% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_get_metadata}
\alias{storage_get_metadata}
\title{Get Storage Object Metadata}
\usage{
storage_get_metadata(conn, object_name, client_id = NULL, client_secret = NULL)
}
\arguments{
\item{conn}{Firebase connection object}

\item{object_name}{Name/path of the object}

\item{client_id}{OAuth client ID}

\item{client_secret}{OAuth client secret}
}
\value{
Object metadata
}
\description{
Retrieves metadata for a storage object.
}
