% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_get_url}
\alias{storage_get_url}
\title{Get Download URL for Storage Object}
\usage{
storage_get_url(conn, object_name, client_id = NULL, client_secret = NULL)
}
\arguments{
\item{conn}{Firebase connection object}

\item{object_name}{Name/path of the object}

\item{client_id}{OAuth client ID}

\item{client_secret}{OAuth client secret}
}
\value{
Download URL string
}
\description{
Generates a download URL for a storage object.
}
\examples{
\dontrun{
url <- storage_get_url(conn, "images/photo.jpg")
if (interactive()) {
  utils::browseURL(url)
}
}
}
