% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_list}
\alias{storage_list}
\title{List Files in Storage}
\usage{
storage_list(
  conn,
  prefix = NULL,
  delimiter = NULL,
  max_results = 1000,
  client_id = NULL,
  client_secret = NULL
)
}
\arguments{
\item{conn}{Firebase connection object}

\item{prefix}{Filter results to objects with this prefix}

\item{delimiter}{Use "/" to list only immediate children}

\item{max_results}{Maximum number of results}

\item{client_id}{OAuth client ID}

\item{client_secret}{OAuth client secret}
}
\value{
List of storage objects
}
\description{
Lists all files in a storage bucket or prefix.
}
\examples{
\dontrun{
# List all files
files <- storage_list(conn)

# List files in a "folder"
images <- storage_list(conn, prefix = "images/")
}
}
