% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_upload_folder}
\alias{storage_upload_folder}
\title{Upload Folder to Storage}
\usage{
storage_upload_folder(
  conn,
  folder_path,
  prefix = "",
  client_id = NULL,
  client_secret = NULL
)
}
\arguments{
\item{conn}{Firebase connection object}

\item{folder_path}{Local folder path}

\item{prefix}{Storage prefix/path for uploaded files}

\item{client_id}{OAuth client ID}

\item{client_secret}{OAuth client secret}
}
\value{
List of upload results
}
\description{
Uploads all files from a local folder to storage.
}
\examples{
\dontrun{
results <- storage_upload_folder(conn, "local/images", "images")
}
}
