% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{upload_folder}
\alias{upload_folder}
\title{This function uploads a local folder to the storage bucket::}
\usage{
upload_folder(bucket_name, web_client_id, web_client_secret, folder_path)
}
\arguments{
\item{bucket_name}{The name of your storage bucket. {string}}

\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}

\item{folder_path}{The path to the folder on the local machine. {string}}
}
\value{
Returns list of storage object ressources of all the files uploaded in the bucket.
}
\description{
fireData::upload_folder uploads all files in the folder to the firebase storage bucket.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
