% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{upload_storage}
\alias{upload_storage}
\title{Upload to Storage (Legacy)}
\usage{
upload_storage(
  bucket_name,
  object_name,
  web_client_id = "prompt",
  web_client_secret = "prompt",
  file_path = NULL,
  predefined_acl = "publicRead"
)
}
\arguments{
\item{bucket_name}{Storage bucket name}

\item{object_name}{Object name in storage}

\item{web_client_id}{OAuth client ID}

\item{web_client_secret}{OAuth client secret}

\item{file_path}{Local file path}

\item{predefined_acl}{Access control}
}
\value{
Storage object metadata
}
\description{
Legacy function. Use storage_upload() instead.
}
