% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{anonymous_login}
\alias{anonymous_login}
\title{The anonymous login function:}
\usage{
anonymous_login(project_api)
}
\arguments{
\item{project_api}{The Firebase Project API {string}}
}
\value{
Returns the content of the firebase API request, such as the idToken, the refreshToken, and the localId.
}
\description{
fireData::anonymous_login signs in a user anonymously.
}
\examples{
\dontrun{
anonymous_login(project_api = AIzaSyAYX_RxenP08jWpHExzsZIaO3CXeZhUSW8)
}
}
\author{
Paul Spende
}
