% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{auth}
\alias{auth}
\title{The user authentication function:}
\usage{
auth(projectAPI, email, password)
}
\arguments{
\item{projectAPI}{The Firebase Project API {string}}

\item{email}{The user email {string}}

\item{password}{The user password {string}}
}
\value{
Returns the content of the firebase API request, such as the state of registration, idToken, and validity of the user password.
}
\description{
fireData_auth checks the validity of a login and returns the temporary JWT user token. FireData_auth can be used to store individual user data in specified directories that are only accessible to that specific user.
}
