% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{classConversion}
\alias{classConversion}
\title{Internal class to binary conversion:}
\usage{
classConversion(x)
}
\arguments{
\item{x}{is the S4 class object}
}
\value{
returns base64 encoded binary value of class object
}
\description{
The conversion is needed to conserve all class information
}
