% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{createUser}
\alias{createUser}
\title{Create User (Legacy)}
\usage{
createUser(projectAPI, email = "prompt", password = "prompt")
}
\arguments{
\item{projectAPI}{The Firebase Project API key}

\item{email}{User email}

\item{password}{User password}
}
\value{
Registration response
}
\description{
Legacy user creation. Use auth_create_user() instead.
}
