% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{createUser}
\alias{createUser}
\title{Firebase user creation function}
\usage{
createUser(projectAPI, email, password)
}
\arguments{
\item{projectAPI}{The Firebase Project API {string}}

\item{email}{The user email {string}}

\item{password}{The user password {string}}
}
\value{
Registers a new user and returns the status.
}
\description{
Creates a new firebase user account. All user accounts can accessed at the firebase.com project console. One of the advantages of firebase accounts in R is the ability to access a website and analyse the data of the website with the very same login.
}
