% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{dataBackup}
\alias{dataBackup}
\title{Database Backup (Legacy)}
\usage{
dataBackup(projectURL, secretKey = "prompt", fileName)
}
\arguments{
\item{projectURL}{Firebase database URL}

\item{secretKey}{Admin secret key}

\item{fileName}{Output file name}
}
\value{
File name
}
\description{
Legacy backup function. Use rtdb_backup() instead.
}
