% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{DataBackup}
\alias{DataBackup}
\title{FireData_backup()}
\usage{
DataBackup(projectURL, secretKey, fileName)
}
\arguments{
\item{secretKey}{The firebase secret key, which can be found in the Config/ Service Accounts/ Database secrets firebase page. {string}}

\item{fileName}{The output file name. Can be any string with .json format {string}}

\item{projectUrl}{The Firebase Project Url {string}}
}
\value{
Returns either a warning or the backup file name.
}
