% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{delete}
\alias{delete}
\title{The firebase data delete function:}
\usage{
delete(x, projectURL, directory = "main", token = "none")
}
\arguments{
\item{x}{A list, data.frame or data.table {object}}

\item{projectURL}{The Firebase project URL {string}}

\item{directory}{The optimal Firebase subdirectory {string}}

\item{token}{The user access token that can be retrieved with the auth() function. Required when if the database rules specify the need for user authentications. {string}}
}
\value{
returns http request answer.
}
\description{
The function allows to delete data objects, such as variables, lists and data.frames
}
\examples{
\dontrun{
delete(x = mtcars, projectURL = "https://firedata-b0e54.firebaseio.com/", directory = "main")
}
}
\author{
Kevin McGinley
}
