% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{download}
\alias{download}
\title{Data conversion function}
\usage{
download(projectURL, fileName, secretKey = "none", token = "none",
  isClass = FALSE)
}
\arguments{
\item{projectURL}{The firebase database url. {string}}

\item{fileName}{The filename or subdirectory. {string}}

\item{secretKey}{The optional database secret key for admin access. {string}}

\item{token}{The user access token that can be retrieved with the auth() function. Required when if the database rules specify the need for user authentications. {string}}

\item{isClass}{In case a s4 class object is downloaded, fireData expects a isClass=TRUE}
}
\value{
returns optionally reformatted data.
}
\description{
The internal data conversion function to bring data in the right json format. In case the uploaded file is a s4 class object, the object is converted to a binary s4 object.
}
