% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{download}
\alias{download}
\title{The firebase data download function:}
\usage{
download(projectURL, fileName, secretKey = "none")
}
\arguments{
\item{x}{A data.frame or data.table}
}
\value{
showing shapiro.test output of the data.frame
}
\description{
fireData_download is the packages main data download function. The function provides the neccessary boilerplate code to load firebase realtime database data into R sessions. Those data can be variables, data.frames, lists and even s4 classes.
}
