% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-links.R
\name{firebase_dynamic_links}
\alias{firebase_dynamic_links}
\title{Firebase Dynamic Links Service}
\description{
Functions for creating Firebase Dynamic Links (short URLs)
}
\section{Note}{

Firebase Dynamic Links is deprecated and will be shut down in August 2025.
Consider using alternative URL shortening services for new projects.
}

