% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_dynamic_link}
\alias{get_dynamic_link}
\title{This function creates short dynamic links::}
\usage{
get_dynamic_link(project_api, long_url)
}
\arguments{
\item{project_api}{The Firebase Project API {string}}
}
\value{
Returns json object with short url and further information.
}
\description{
fireData::get_dynamic_link creates a short link for a longer one.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
