% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_dynamic_link}
\alias{get_dynamic_link}
\title{This function creates short dynamic links::}
\usage{
get_dynamic_link(project_api, domain, long_url, option = "SHORT")
}
\arguments{
\item{project_api}{The Firebase Project API {string}}

\item{domain}{The Firebase dynamic link domain {string}}

\item{long_url}{The URL you want to shorten {string}}

\item{option}{The url generation option ("SHORT" or "UNGUESSABLE"). {string}}
}
\value{
Returns json object with short url and further information.
}
\description{
fireData::get_dynamic_link creates a short link for a longer one.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
