% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_storage}
\alias{get_storage}
\title{The get function for firebase storage:}
\usage{
get_storage(bucket_name, object_name, web_client_id = "prompt",
  web_client_secret = "prompt")
}
\arguments{
\item{bucket_name}{The name of your storage bucket. {string}}

\item{object_name}{The name of the file that you want to get from the bucket. {string}}

\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}
}
\value{
Returns the storage object ressource for the file in the bucket.
}
\description{
fireData::get_storage lists all files in the firebase storage bucket.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
