% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_url}
\alias{get_url}
\title{The get the url for an object in storage (e.g. an htmk file):}
\usage{
get_url(bucket_name, object_name, web_client_id = "prompt",
  web_client_secret = "prompt")
}
\arguments{
\item{bucket_name}{The name of your storage bucket. {string}}

\item{object_name}{The name of the file that you want to get from the bucket. {string}}

\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}
}
\value{
Returns the url for the storage object.
}
\description{
fireData::get_url outputs the url to access the object in storage.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
