% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{google_firestore}
\alias{google_firestore}
\title{The OAuth function to get access to the cloud firestore:}
\usage{
google_firestore(web_client_id = "prompt", web_client_secret = "prompt")
}
\arguments{
\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}
}
\value{
Returns the token data.
}
\description{
fireData::google_firestore retrieves a token with read access to the cloud firestore
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
