% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{google_login}
\alias{google_login}
\title{Google Login (Legacy)}
\usage{
google_login(
  project_api,
  web_client_id = "prompt",
  web_client_secret = "prompt",
  request_uri,
  redirect_uri = httr::oauth_callback(),
  return_idp_credential = TRUE,
  cache = FALSE
)
}
\arguments{
\item{project_api}{Firebase API key}

\item{web_client_id}{Google OAuth client ID}

\item{web_client_secret}{Google OAuth client secret}

\item{request_uri}{Request URI}

\item{redirect_uri}{Redirect URI}

\item{return_idp_credential}{Whether to return IDP credential}

\item{cache}{Whether to cache token}
}
\value{
Authentication response
}
\description{
Legacy Google login. Use auth_google() instead.
}
