% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{o_auth_login}
\alias{o_auth_login}
\title{OAuth Login (Legacy)}
\usage{
o_auth_login(project_api, request_uri, post_body, return_idp_credential)
}
\arguments{
\item{project_api}{Firebase API key}

\item{request_uri}{Request URI}

\item{post_body}{OAuth post body}

\item{return_idp_credential}{Whether to return IDP credential}
}
\value{
Authentication response
}
\description{
Legacy OAuth login. Use auth_oauth() instead.
}
