% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{resetPassword}
\alias{resetPassword}
\title{Password resett function:}
\usage{
resetPassword(projectAPI, email)
}
\arguments{
\item{projectAPI}{The Firebase Project API {string}}

\item{email}{The user email {string}}
}
\value{
Success or warning message.
}
\description{
Resets the user password and sends an email to the user account.
}
