% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{resetPassword}
\alias{resetPassword}
\title{Reset Password (Legacy)}
\usage{
resetPassword(projectAPI, email)
}
\arguments{
\item{projectAPI}{The Firebase Project API key}

\item{email}{User email}
}
\value{
Success or warning message
}
\description{
Legacy password reset. Use auth_reset_password() instead.
}
