% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R
\name{ServiceAccountCredentials}
\alias{ServiceAccountCredentials}
\title{Service Account Credentials}
\description{
Represents credentials from a Google service account.
}
\keyword{Print}
\keyword{credentials}
\keyword{internal}
\keyword{summary}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{project_id}}{GCP project ID from service account}

\item{\code{client_email}}{Service account email address}

\item{\code{private_key}}{RSA private key for signing JWTs}

\item{\code{access_token}}{Current OAuth access token}

\item{\code{token_expires_at}}{Access token expiration time
Create from service account JSON file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ServiceAccountCredentials-new}{\code{ServiceAccountCredentials$new()}}
\item \href{#method-ServiceAccountCredentials-get_access_token}{\code{ServiceAccountCredentials$get_access_token()}}
\item \href{#method-ServiceAccountCredentials-create_jwt}{\code{ServiceAccountCredentials$create_jwt()}}
\item \href{#method-ServiceAccountCredentials-print}{\code{ServiceAccountCredentials$print()}}
\item \href{#method-ServiceAccountCredentials-clone}{\code{ServiceAccountCredentials$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceAccountCredentials-new"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceAccountCredentials-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceAccountCredentials$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to service account JSON file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ServiceAccountCredentials object
Get an access token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceAccountCredentials-get_access_token"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceAccountCredentials-get_access_token}{}}}
\subsection{Method \code{get_access_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceAccountCredentials$get_access_token(
  scopes = c("https://www.googleapis.com/auth/firebase.database",
    "https://www.googleapis.com/auth/datastore",
    "https://www.googleapis.com/auth/devstorage.read_write")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scopes}}{OAuth2 scopes to request}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Access token string
Create a signed JWT
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceAccountCredentials-create_jwt"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceAccountCredentials-create_jwt}{}}}
\subsection{Method \code{create_jwt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceAccountCredentials$create_jwt(scopes, lifetime = 3600)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scopes}}{OAuth2 scopes}

\item{\code{lifetime}}{Token lifetime in seconds}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Signed JWT string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceAccountCredentials-print"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceAccountCredentials-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceAccountCredentials$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceAccountCredentials-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceAccountCredentials-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceAccountCredentials$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
