% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{shiny_auth_server}
\alias{shiny_auth_server}
\title{This function creates a login overlay for a shiny application::}
\usage{
shiny_auth_server(user, input, project_api, web_client_id,
  web_client_secret, request_uri)
}
\arguments{
\item{user}{A flag if user is logged in or not. E.g.   USER <- reactiveValues(Logged = FALSE) {reactiveValues}}

\item{input}{The input to the shiny server. {input}}

\item{project_api}{The Firebase Project API {string}}

\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}

\item{request_uri}{The URI to which the IDP redirects the user back. {string}}
}
\value{
Returns a functional login overlay.
}
\description{
fireData::shiny_auth creates a short link for a longer one.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
