% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realtime-db.R
\name{upload}
\alias{upload}
\title{Upload Data (Legacy)}
\usage{
upload(x, projectURL, directory = "main", token = "none")
}
\arguments{
\item{x}{Data to upload}

\item{projectURL}{Firebase database URL}

\item{directory}{Database path}

\item{token}{Authentication token}
}
\value{
Path with generated key
}
\description{
Legacy upload function. Use rtdb_push() instead.
}
