% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{upload}
\alias{upload}
\title{The firebase data upload function:}
\usage{
upload(x, projectURL, directory = "main")
}
\arguments{
\item{x}{A data.frame or data.table {object}}

\item{projectURL}{The Firebase project URL {string}}

\item{directory}{The optimal Firebase subdirectory {string}}
}
\value{
returns http request answer which includes the random key.
}
\description{
The function allows to upload data objects, such as variables,lists and data.frames
}
\author{
Robin Kohze
}
