% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{upload_folder}
\alias{upload_folder}
\title{Upload Folder (Legacy)}
\usage{
upload_folder(
  bucket_name,
  web_client_id = "prompt",
  web_client_secret = "prompt",
  folder_path
)
}
\arguments{
\item{bucket_name}{Storage bucket name}

\item{web_client_id}{OAuth client ID}

\item{web_client_secret}{OAuth client secret}

\item{folder_path}{Local folder path}
}
\value{
List of upload results
}
\description{
Legacy function. Use storage_upload_folder() instead.
}
