% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{upload_storage}
\alias{upload_storage}
\title{The upload function for firebase storage:}
\usage{
upload_storage(bucket_name, object_name, web_client_id, web_client_secret,
  file_path = NULL)
}
\arguments{
\item{bucket_name}{The name of your storage bucket. {string}}

\item{object_name}{The name you want to give your file in the storage. {string}}

\item{web_client_id}{The Web Client ID of your Google OAuth in your Firebase. {string}}

\item{web_client_secret}{The Web Client Secret of your Google OAuth in your Firebase. {string}}

\item{file_path}{The path of the file you want to upload, leave this empty to create a folder. {string}}
}
\value{
Returns the storage object informations.
}
\description{
fireData::upload_storage uploads a file to the firebase storage.
}
\examples{
\dontrun{
TODO:
}
}
\author{
Paul Spende
}
